* p4a-packages:
  [Pointer Values] take into account unification of pointer values null value with points-to lattice null_pointer location
  [Pointer Values and effects] generic effects aliases search with simple pointer values
  Remove debug (wrongly commited)
  [Liveness] new test case for liveness with pointer aliasing
  [Liveness] Correct results with constant path effects without pointer analysis or with pointer values
  Add example that show how pointer dereferencing break liveness analysis
  Declarations with initialization break perfect nesting of loops
  Illustrate the bad behavior of depth_of_parallel_perfect_loop_nest() in presence of declaration with initialization
  [Semantics] remove compilation warnings
  Pointers: missing tpips scripts and some bugs in C code; update wrt new test cases added by Amira on Friday
  alias-classes: passes.c slightly updated; hopefully, the initial version of points-to analysis is going to fare better; previous issue was linked to the interface witj EffectsWithPointsTo
  Add an example to enforce SCALARIZATION_KEEP_PERFECT_PARALLEL_LOOP_NESTS property
  Add a property "SCALARIZATION_KEEP_PERFECT_PARALLEL_LOOP_NESTS"
  Use contexts with gen_recurse instead of global variables
  Indentation
  Pointers: updates for the new version
  alias-classes: check-in of new version; not clear how much the initial version is preserved; computation of points-to IN modified; sumarization for loops in trouble; bug in sort fixed
  transformer/io.c: bug fix in fprint_transformer(); the same bug exists in many print_xxx() fuctions I have recently modified to check the type dynamically
  Semantics-New/unisgned08.result: svn add forgotten
  Semantics-New: new test cases and results related to type information and modulo
  semantics/ri_to_transformers.c: reverse functionally previous commit; 1) the bug was in the tpips script, 2) the previous commit was an unfortunate control-P/return
  semantics/ri_to_transformers.c, ri_to_preconditions.c: preliminary handling of type information added
  semantics/ri_to_transformers.c, ri_to_preconditions.c: preliminary handling of type information added
  effects-util/points_to.c: two new predicate function to detect cells related to formal parameter or points-to stubs
  transformer: dynamic type checking for debug function; constant value changed from  long int into lon long int; this type chagne should be performed on many more transformer constructor or update functions
  ri-util/expression.c: print_expression() restructured to make it easier to cut-and-paste the dynamic type checking and the test to expression_undefined
  pipsmake-rc.tex: new property added to use type information in semantics analysis
  Semantics-New/modulo10: refine transformers required to exploit interprocedural information when linearizing non-linear operators; IN_CONTEXT option limited to the propagation of intraprocedural information
  Add another example for #697 'Precondition is not precise enough after unrolling (modulo inside)'
  Semantics-New/modulo06: transformer refinement added as well as correct result
  Semantics-New/modulo09: property SEMANTICS_COMPUTE_TRANSFORMERS_IN_CONTEXT must be set to true or transformers be refined to obtain information about modulo operation; Ticket 697
  Add example for #697 : "Precondition is not precise enough after unrolling (modulo inside)"
  Add examples for #696 : "Unsigned is not taken into account by preconditions"
  Add example for #695 "Preconditions for modulo is less precise on formal parameters"
  Add missing file linked_list01.c and .tpips
  Add missing files from EffectsWithPointsTo; maybe some cleaning will be necessery
  [Pointer Values and regions] new interface to compute regions using pointer values
  Pointers: lots of small new test cases; tpips scripts added; sometimes, test files fixed, which generates 40 differences for the initial version of points-to
  alias-classes: another check point
  [Effects with Pointer Values] some more test cases for pointer arithmetic and dereferencing undefined pointers
  [PointerValues] change behavior in case of undefined or null pointer value (r21200)
  [Pointer Values and Effects] take into account null and undefined pointer values
  [ri-util/expression_to_type] fix type of p-q when p and q are pointers, according to c99 standard (returned type was pointer instead of ptrdiff_t)
  [effects_util/prettyprint] fix compilation errors due to non matching qualifiers
  More Emami basic cases
  Pointers: updates for reengineered version of points-to
  alias-classes: check point of new version; about done with the Pointers validation suite
  Add Emami basic  cases
  alias-classes: another checkpoint for the reengineered version
  Pointers: updates mostly; sorting issue improved; still a potential problem for entities that are not program variables but abstract locations; the number of problems for the current vetrsion increases, but this seems to be due to the new test cases recently introduced for debugging the new version
  Do not activate PHPiPS by default, has to be activated by --enable-phpips
  Add the first release for the brand new PHPiPS interface!!
  ri-util: predicate static_global_variable_p() added
  Pointers: new test cases for the new implementation of points-to; added tpips scripts used for debuggin; test files for new test cases added consistent with new implementation
  alias-classes: checkpoint for the new version of the points-to analysis
  Fix compilation in devel-mode
  Pointers: global07 added
  alias-classes: another checkpoint
  Pointers: new test cases for new version + more tpips scripts + a fix for array04.result/test
  alias-classes: check-point for the new version; bug fix for naming of stub variables associated to global variables; name changes for function used in effects-util, requested by Beatrice, not performed yet; number of validation problems increased... because the number of test cases has increased too and some of them are designed for the new version
  alias-classes: unification of inter and intra-procedural points-to analysis; function interproceudral_points_to_analysis_p() added for test in user call handling


Former-commit-id: e40320c45cb11dde26e97d92acda3804c4ee55c2

+++ packages/PIPS/pips/src/Libs/transformer/basic.c
@@ -424,7 +424,7 @@ transformer transformer_add_equality(transformer tf, entity v1, entity v2)
 }
 
 /* Add an equality between a value and an integer constant: v==cst */
-transformer transformer_add_equality_with_integer_constant(transformer tf, entity v, int cst)
+transformer transformer_add_equality_with_integer_constant(transformer tf, entity v, long long int cst)
 {
   Pvecteur eq = vect_new((Variable) v, VALUE_MONE);
 
@@ -451,7 +451,7 @@ transformer transformer_add_inequality(transformer tf, entity v1, entity v2, boo
 }
 
 /* Add the inequality v <= cst or v >= cst */
-transformer transformer_add_inequality_with_integer_constraint(transformer tf, entity v, int cst, bool less_than_p)
+transformer transformer_add_inequality_with_integer_constraint(transformer tf, entity v, long long int cst, bool less_than_p)
 {
   Pvecteur eq = vect_new((Variable) v, VALUE_ONE);
